+++ brikar-server/src/main/java/com/truward/brikar/server/tracking/RequestIdAwareFilter.java
@@ -17,6 +17,9 @@ import java.io.IOException;
 /**
  * A filter which performs processing of request ID and originating reqiest ID, such as
  * retrieving and propagating these IDs to the MDC context and putting these IDs to the response headers.
+ *
+ * This filter is associated with logger, which produces results under <code>BrikarRequestLogger</code> name.
+ *
  * See also {@link LogUtil} and {@link TrackingHttpHeaderNames}.
  *
  * @author Alexander Shabanov
@@ -28,12 +31,7 @@ public class RequestIdAwareFilter extends OncePerRequestFilter {
   private static final String OK_LOG_FORMAT = LogUtil.METRIC_HEADING + ", " + LogUtil.VERB + "={}, " +
       LogUtil.TIME_DELTA + "={}, " + LogUtil.RESPONSE_CODE + "={}, " + LogUtil.URL + "={}";
 
-  private final Logger log;
-
-  public RequestIdAwareFilter() {
-    final Logger log = LoggerFactory.getLogger(getClass());
-    this.log = log.isTraceEnabled() ? log : null;
-  }
+  private final Logger log = LoggerFactory.getLogger("BrikarRequestLogger");
 
   @Override
   protected void doFilterInternal(HttpServletRequest request,
@@ -55,9 +53,7 @@ public class RequestIdAwareFilter extends OncePerRequestFilter {
     response.setHeader(TrackingHttpHeaderNames.REQUEST_ID, originatingRequestId);
 
     // process request
-    if (log == null) {
-      filterChain.doFilter(request, response);
-    } else {
+    if (log.isTraceEnabled()) {
       String pathInfo = request.getPathInfo(); // can be null if request doesn't have path info
       if (pathInfo != null) {
         pathInfo = LogUtil.encodeString(pathInfo);
@@ -81,6 +77,8 @@ public class RequestIdAwareFilter extends OncePerRequestFilter {
 
         log.info(OK_LOG_FORMAT, operation.toString(), request.getMethod(), time, response.getStatus(), pathInfo);
       }
+    } else {
+      filterChain.doFilter(request, response);
     }
 
     // remove MDC variables